added cpu util

+++ include/processor.h
@@ -7,7 +7,7 @@ class Processor {
 
   // TODO: Declare any necessary private members
  private:
-    unsigned long long prev_idle{0};
+    long double prev_idle{0};
     unsigned long long prev_total{0};
 };
